requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-openaip
langchain-huggingface
faiss-cpu
python-dotenv
amadeus
requests
pydantic
jq
sentence-transformers
google-search-results
mcp
